requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
async-generator==1.10
attrs==25.4.0
base58==1.0.3
bitarray==3.8.0
blessed==1.24.0
blinker==1.9.0
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
contexttimer==0.3.3
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
dnspython==2.8.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
hexbytes==1.3.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lru-dict==1.2.0
MarkupSafe==3.0.3
matplotlib==3.8.4
morphys==1.0
multiaddr==0.1.1
multidict==6.7.0
mypy_extensions==1.1.0
netaddr==1.3.0
numpy==2.0.1
nvm==1.0.21
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.2
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
py-cid==0.3.1
py-ecc==8.0.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==2.0.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pymultihash==0.8.2
pyparsing==3.2.5
pyrsistent==0.20.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==17.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
srsly==2.5.2
toolz==1.1.0
trio==0.32.0
trio-typing==0.10.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
varint==1.0.2
wcwidth==0.2.14
web3==7.14.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
